refactor: reorganize config structure
