+++ script/Config/Audio/AudioConfig.lua
@@ -1,0 +1,9 @@
+Config.audio.general = {
+    soundEnabled         = true,
+    supportedFormats     = { ".ogg", ".mp3", ".wav", ".flac" },
+    mainMenuMusicEnabled = true,
+    soundMin             = 0,
+    soundMax             = 1,   -- SetVolume range seems to go from 0 (min) to about 2 or 3 (max)
+    musicVolume          = 0.5, -- current volume
+    mainMenu             = "LTR_Explorer.ogg",
+}
