+++ script/Config/General/GameConfig.lua
@@ -1,0 +1,72 @@
+--! this can be split up better
+Config.game = {
+    boostCost              = 20,
+
+    explosionSize          = 64,
+
+    autoTarget             = false,
+
+    pulseColorBodyR        = 0.3,
+    pulseColorBodyG        = 0.8,
+    pulseColorBodyB        = 2.0,
+    pulseColorLightR       = 0.3,
+    pulseColorLightG       = 0.9,
+    pulseColorLightB       = 3.0,
+
+    droneType              = 1, -- 1 = mining drone, 2 = combat drone
+    droneDamage            = 10,
+    droneTarget            = nil,
+    droneSize              = 75,
+    droneSpeed             = 50,
+    droneRange             = 5000,
+
+    missileDamage          = 80,
+    missileTarget          = nil,
+    missileSize            = 100,
+    missileSpeed           = 100,
+    missileRange           = 10000,
+
+    torpedoDamage          = 250,
+    torpedoTarget          = nil,
+    torpedoSize            = 64,
+    torpedoSpeed           = 30,
+    torpedoRange           = 14000,
+
+    weaponGroup            = 1,
+
+    shieldColor            = Color(0.2, 0.5, 1.0, 1.0),
+
+    enemies                = 0,
+    friendlies             = 0,
+    squadSizeEnemy         = 8,
+    squadSizeFriendly      = 8,
+
+    spawnDistance          = 2000,
+    friendlySpawnCount     = 10,
+    timeScaleShipEditor    = 0.0,
+
+    aiFire                 = function(dt, rng) return rng:getExp() ^ 2 < dt end,
+
+    autonavRanges          = { 200, -- Unknown
+        0,                          -- Reserved
+        0,                          -- Star Sector
+        0,                          -- Star System
+        2000,                       -- Zone
+        1e7,                        -- Star (TODO: radius + offset)
+        10000,                      -- Planet (TODO: radius + offset)
+        300,                        -- Asteroid
+        500,                        -- Jumpgate
+        2000,                       -- Station
+        100 },                      -- Ship
+
+    dockRange              = 50,
+
+    dispoMin               = -1.0,
+    dispoNeutral           = 0.0,
+    dispoMax               = 1.0,
+    dispoHostileThreshold  = -0.3333333,
+    dispoFriendlyThreshold = 0.3333333,
+    dispoName              = { "hostile",
+        "neutral",
+        "friendly" },
+}
