+++ script/Config/Render/RenderConfig.lua
@@ -1,0 +1,11 @@
+Config.render = {
+    defaultResX    = 1920,
+    defaultResY    = 1080,
+    fullscreen     = false,
+    presentMode    = PresentMode.AutoVsync,
+    fov            = 70,
+    zNear          = 0.1, -- default: 0.1
+    zFar           = 1e6, -- default: 1e6
+    thrusterLights = false,
+    pulseLights    = false,
+}
