+++ script/Config/UI/FontConfig.lua
@@ -1,0 +1,6 @@
+Config.ui.font = {
+    normal     = Cache.Font('Share', 14),
+    normalSize = 14,
+    title      = Cache.Font('Exo2Bold', 10),
+    titleSize  = 10,
+}
